feat: replace uint16_t file_id with Velox StringIdMap uint64_t
- LiquidCacheKey::file_id widened from uint16_t to uint64_t
- Use fileIds().makeId(path) for deterministic file IDs in Velox mode
- Fall back to uint64_t sequential counter in non-Velox mode
- Hold StringIdLease in LiquidCacheStore for lifecycle management
- Remove to_u64()/from_u64() (112-bit key no longer fits in 64 bits)
- Use field-wise operator== and hash combination

+++ src/liquid_to_velox.cpp
@@ -670,7 +670,7 @@ VectorPtr LiquidCacheStore::read_column_velox(
 // ═══════════════════════════════════════════════════════════════════════
 
 VectorPtr LiquidCacheStore::read_batch_velox(
-        uint16_t file_id,
+        uint64_t file_id,
         uint16_t rg_id,
         uint16_t batch_id,
         const RowTypePtr& rowType,
